api/requirements.txt
Django==3.2.22
psycopg2-binary==2.8.6
gunicorn==20.1.0
bcrypt==3.2.0
djangorestframework-simplejwt==4.1.2
django-cors-headers==3.7.0
pillow==9.3.0
Faker==8.1.1
nplusone==1.0.0
pytest==6.2.4
factory_boy==3.2.0
python-dotenv==0.17.1
pytz==2021.1
djangorestframework==3.12.4
pytest-factoryboy==2.5.1
djoser==2.0.5
pytest-django==4.4.0
django-filter==23.2
